requirements.txt
aiohttp==3.11.16
datasets==3.5.0
decord==0.6.0
matplotlib==3.10.3
numpy==2.2.6
openai==1.82.0
pandas==2.2.3
Pillow==11.2.1
protobuf==6.31.0
pycocotools==2.0.8
qwen_vl_utils==0.0.11
Requests==2.32.3
tenacity==9.1.2
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
# transformers==5.1.0
transformers==4.57.0
openpyxl
opencv-python-headless
accelerate
nltk
sentence-transformers==5.1.2
spacy
keye-vl-utils==1.5.2